Fix ineffectual assignments
This is a tougher PR to review, I'm sorry. I fully understand if you reject it
and approach this from a different angle. However I've stumbled upon a whole
range of ineffectual assignments in your code.

Most of them were assignments to an error variable, but missing the actual error
check. In the various test cases, I've tried to stick to whatever was the convention
in the particular test: require.NoError, require.Nil, direct err check.

I suggest you want to annotate a bunch of them. There were too many here for me
to do this in a reasonable fashion in this PR.

Should you intentionally skip these checks, I would recommend omitting the assignment
as well, as this clearly states an intention to future readers/reviewers.