fix: wrong history cache test
The test was checking for failure if the directory does not exist,
but the code mkdirs the directories it needs.

The test passed on Linux by accident: the non-existing directory
used in the test was /not-exists and failed not because of
non-existence but because of permissions.

There is a chance the test was correct, and actually testing just
fishing coverage points for the rare case of permissions.
If this is the case, the test did not work on Windows, because admin
can anyway create the directory.