fix: tlsConfig is always non-nil
Before pgsql branch merge, tlsConfig was nil if no TLS configured.

We now do not make assumptions about the resulting tlsConfig object.
We do only start a TLS server if tlsConfig is non-nil and has certificates,
but we only test the resulting server can accept the right type of connection.