chore(Makefile): More explicit webconsole version
Makefile now contains an information what webconsole version
is the best match for the current immudb code. That version is
selected when building with `make WEBSONCOLE=default ...`.
Alternatively one can select the latest released version of
webconsole with `make WEBCONSOLE=latest ...` or use a specific
version with `make WEBCONSOLE=<version> ...`.

Signed-off-by: Bartłomiej Święcki <bart@codenotary.com>