use sql.TimeToInt64 for converting time to an sql timestamp
This fixes a bug with timestamp columns, which were not read and written
corrently if the client inserts or queries a timestamp column. The db
engine stores TIMESTAMP values in microseconds, but the client was
sending nanoseconds. This caused wrong values being put in
TIMESTAMP columns. The queried TIMESTAMP values were also send in
nanoseconds to the client. To improve consistency, these values were also changed
to microseconds.