fix(embedded/store): Protect against simultaneous replicators
The check for expected transaction header is made without holding
the transaction lock. This would work as long as there is only
one writer in such case which is what happens in a standard
setup. In case of embedded though this is not guaranteed and there
could be simultaneous writers for replica (and other use cases
passing the expected header) thus we need to ensure that validation
still holds while the final commit mutex lock is held.

Signed-off-by: Bartłomiej Święcki <bart@codenotary.com>