test(pkg/integration/replicatino): Add WaitForCommittedTx
This method is a useful utility to ensure that the
replica committed given transaction id. This is necessary
because shortly after replica confirms the write to primary
that transaction will still be in a precommitted state thus
we have to wait for another sync cycle on replica to make it
committed.

Signed-off-by: Bartłomiej Święcki <bart@codenotary.com>