fix(verification): Additional Linear proof consistency check
Require additional proof that the linear proof part that was previously
accepted by the client and that was consumed by the new Merkle Tree
is stored in leaf nodes of that new Merkle Tree.

Without this check, the server can try to fool the client by
jumping over multiple nodes in the linear chain where only the
first and the last element is really checked giving it an opportunity
to swap previously accepted transaction with a modified one.

This change is part of CVE-2022-36111 fix.

Signed-off-by: Bartłomiej Święcki <bart@codenotary.com>