fix(verification): Recreate linear advance proofs for older servers
In case of older servers that do not generate linear advance proofs,
those proofs are manually filled in by the client by additional calls
to the server.

The number of additional calls to the sever is proportional to the
length of the linear chain to prove. This should be small in
real-world scenarios and could only happen in immudb servers before
1.3.2 version in a concurrent write scenarios. In case of such scenarios
though, verified operations may be much slower compared to previous
immudb versions.

This change is part of CVE-2022-36111 fix.

Signed-off-by: Bartłomiej Święcki <bart@codenotary.com>