fix: time.Since should not be used in defer statement
This commit fixes the complaining from `go vet` about the deferring of
`time.Since`. This can cause wrong time evaluation of the elapsed time.
This impact only the debug level, but still it should be addressed.

To make the code a little bit more readable, I wrote a method on the
`immuClient` instance with the signature `debugElapsedTime(method
string, start time.Time)`.

ref: https://github.com/golang/go/issues/60048

Signed-off-by: Maurizio Del Corno <druznek@gmail.com>